feature/refactor-core-fb-js
REFACTOR `form-builder.js` to use ES6 Class and be more modular.

- Previous setup with IIFE could not be easily expanded.
- Structure of the file used a lot of jQuery `$.extend`.
- Layout for `form_div` was missing (only Bootstrap v3 and v4).

In my current client/project we needed to use the "form_div_layout"
since we are using Foundation for our frontend. The case was missing in
the plugin and had to be extended.

I used the opportunity to refactor the code and be easier to debug. With
this change the whole core plugin should be more modular and easier
to read.